Default implementations in AbstractCoordinate
AbstractCoordinate now has default implementations for the methods getCartesianDistance,
getCentralAngle and isEqual.

This is achieved by moving up the same named method from the subclass in the super class.
In AbstractCoordinate the coordinates are converted to CartesianCoordinate or SphericCoordinate
inside the methods. For convertion the interface Coordinate already defines
asCartesianCoordinate and asSphericCoordinate.

To access the private attributes, public getter-methods were introduced in the
subclasses.